make updateUi independ
