+++ src/app/CMakeLists.txt
@@ -1,4 +1,4 @@
-add_executable(app ./main.cpp ./app.cpp ./callbacks.cpp)
+add_executable(app ./main.cpp ./app.cpp ./callbacks.cpp ./updateUi.cpp)
 target_compile_features(app PRIVATE cxx_std_17)
 
 add_subdirectory(./context)
